Add BoardRepository

+++ src/main/kotlin/mn/kotlin/domain/Board.kt
@@ -1,3 +1,3 @@
 package mn.kotlin.domain
 
-data class Board(val id: Int?, val name: String, val description: String)
+data class Board(val id: Long?, val name: String, val description: String)
